Angular Lighting Incentives Search - Versions A, B and C
This includes 3 versions of my Lighting Incentives Search page using
AngularJS, PHP and MySQL. The first version of the Incentives Search
does not include the prerequisites for each lighting incentive.  That
required some creating some nested JSON to populate the prerequisites
in the later versions.  Also, in the first version of the Incentive
Select PHP file, I used a lengthy “If Else” statement to populate the
images for each lighting type.  In the later versions I replaced the
“If Else” with a “Switch Case”